microservices/orchestrator/pyproject.toml
[tool.poetry]
name = "orchestrator"
version = "0.1.0"
description = ""
authors = ["tropicbliss <48951297+tropicbliss@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.94.1"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
aiohttp = "^3.8.4"
aio-pika = "^9.0.5"
aiographql-client = "^1.0.3"


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.0"
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
